projeto_ia_puc/requirements.txt
streamlit
requests
pyyaml
pdfplumber
easyocr
pdf2image
pillow
torch
numpy